+++ src/java/com/modp/random/BlumBlumShub.java
@@ -201,10 +201,10 @@ public class BlumBlumShub implements RandomGenerator {
      *
      * @return int
      */
-    public int next(int numBits) {
+    public long next(int numBits) {
 	// TODO: find out how many LSB one can extract per cycle.
 	//   it is more than one.
-	int result = 0;
+	long result = 0;
 	for (int i = numBits; i != 0; --i) {
 	    state = state.modPow(two, n);
 	    result = (result << 1) | (state.testBit(0) == true ? 1 : 0);
